requirements.txt
annotated-types==0.7.0
comtypes==1.4.10
MouseInfo==0.1.3
psutil==7.0.0
PyAutoGUI==0.9.54
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==310
pywinauto==0.6.9
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
fastapi~=0.115.12
uvicorn[standard]~=0.34.1
starlette~=0.46.2
pymongo~=4.12.0
Flask~=3.1.0
PyJWT~=2.10.1
motor~=3.7.0
pytest~=8.3.5
httpx~=0.28.1
